ImageFilter tests

+++ src/filter_kernel.h
@@ -8,8 +8,8 @@ class FilterKernel {
 
         virtual std::vector<float> getValues() = 0;
         virtual unsigned short getSize() = 0;
-        virtual float at(size_t col, size_t row) = 0;
-        virtual float at1D(size_t idx) = 0;
+        virtual float at(int col, int row) = 0;
+        virtual float at1D(int idx) = 0;
         virtual bool isSeparableInto1D() = 0;
 };
